pyproject.toml
[project]
name = "device-control"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "google-genai>=1.26.0",
    "loguru>=0.7.3",
    "mem0ai>=0.1.116",
    "objectbox>=4.0.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "statistics>=1.0.3.5",
    "uvicorn>=0.35.0",
]
